requirements.txt
asgiref==3.8.1
crispy-bootstrap5==2025.4
Django==5.2
django-crispy-forms==2.4
django-filter==25.1
sqlparse==0.5.3
typing_extensions==4.13.2
